Consolidate Debian/RedHat service configurations.
Since we're dropping support for anything old enough to not use systemd,
we can consolidate the defaults/sysconfig files - they're both simply
used as EnvironmentFiles in the libvirtd.service unit file, so they can
safely have the same content.

The only exception for this is Ubuntu prior to Jammy - in that case the
defaults file still uses the old 'libvirtd_opts' environment variable,
rather than the 'LIBVIRTD_ARGS' which appears to be the upstream
version.